requirements.txt
requests
beautifulsoup4
lxml
pydantic
pyyaml
python-dateutil